Add extra data to auto network selection message.
If device is already in auto mode, and choose network automatically
is chosen, then network selection complete message is sent to the
application, which triggered the request, without sending the request
to RIL. On the app side, it is not known whether the request was
actually sent to RIL or not.

Add information to the message so that application knows if the request
has been sent to RIL or not.

Change-Id: I33401a93958864ccee2983ffe607dfff7a09154e
CRs-Fixed: 2015326